requirements.txt
ttkbootstrap
pyautogui
Pillow
PyGetWindow
opencv-python
numpy
pywin32
requests
pytest
